fix(rmm): fix MISRA C:2012 Rule 10.4

+++ lib/xlat/src/xlat_tables_utils.c
@@ -317,9 +317,9 @@ static uint64_t *find_xlat_last_table(uintptr_t va,
 	ret_table = ctx_tbls->tables;
 	table_base_va = ctx_cfg->base_va;
 
-	for (int level = start_level; level <= XLAT_TABLE_LEVEL_MAX; level++) {
-		unsigned int idx =
-			(unsigned int)XLAT_TABLE_IDX(va_offset, level);
+	for (int level = start_level;
+	     level <= XLAT_TABLE_LEVEL_MAX; level++) {
+		unsigned int idx = (unsigned int)XLAT_TABLE_IDX(va_offset, level);
 		uint64_t desc = ret_table[idx];
 		uint64_t desc_type = desc & DESC_MASK;
 
@@ -433,7 +433,7 @@ int xlat_map_memory_page_with_attrs(const struct xlat_llt_info * const table,
 	}
 
 	/* Generate the new descriptor */
-	tte = xlat_desc(attrs, (pa & XLAT_ADDR_MASK(table->level)),
+	tte = xlat_desc(attrs, pa & XLAT_ADDR_MASK((unsigned int)table->level),
 			table->level) | TRANSIENT_DESC;
 
 	xlat_write_tte(tte_ptr, tte);
@@ -511,7 +511,7 @@ uint64_t *xlat_get_tte_ptr(const struct xlat_llt_info * const llt,
 	}
 
 	offset = va - llt->llt_base_va;
-	index = (unsigned int)(offset >> XLAT_ADDR_SHIFT(llt->level));
+	index = (unsigned int)(offset >> XLAT_ADDR_SHIFT((unsigned int)llt->level));
 
 	return (index < XLAT_GET_TABLE_ENTRIES(llt->level)) ?
 			&llt->table[index] : NULL;
